requirements.txt
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
audioread==3.0.1
backcall==0.2.0
basic-pitch @ file:///Users/elihu.sela/Dev/HarmonicaTabs/basic-pitch
beautifulsoup4==4.13.4
bleach==6.2.0
cattrs==24.1.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
coremltools==8.3.0
cycler==0.12.1
decorator==5.2.1
defusedxml==0.7.1
docopt==0.6.2
executing==2.2.0
fastjsonschema==2.21.1
fonttools==4.57.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
iniconfig==2.1.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mido==1.3.3
mir_eval==0.8.2
mistune==3.1.3
moviepy==2.1.2
mpmath==1.3.0
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
numba==0.61.2
numpy==2.2.5
packaging==25.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pipreqs==0.5.0
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
pretty_midi==0.2.10
proglog==0.1.11
prompt_toolkit==3.0.51
protobuf==6.30.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==25.1.0
pycparser==2.22
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
resampy==0.4.2
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
soundfile==0.13.1
soupsieve==2.7
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
